refactor: isolate scene digest envelope handling
