+++ python/dcc_mcp_core/script_execution.py
@@ -37,15 +37,14 @@ from dcc_mcp_core.runtime.scene_digest import SceneDigestExecution
 from dcc_mcp_core.runtime.scene_digest import SceneDigestExecutionError
 from dcc_mcp_core.runtime.scene_digest import SceneDigestSnapshot
 from dcc_mcp_core.runtime.scene_digest import StateDigestProvider
-from dcc_mcp_core.runtime.scene_digest import normalize_scene_digest
 from dcc_mcp_core.runtime.scene_digest import snapshot_from_provider
+from dcc_mcp_core.runtime.scene_digest_envelope import scene_digest_postcondition as _scene_digest_postcondition
 from dcc_mcp_core.schema import derive_script_parameters_schema
 from dcc_mcp_core.script_materialization import MaterializedScript
 from dcc_mcp_core.script_materialization import cleanup_materialized_scripts
 from dcc_mcp_core.script_materialization import default_script_materialization_root
 from dcc_mcp_core.script_materialization import materialize_script
 from dcc_mcp_core.script_materialization import resolve_materialized_script
-from dcc_mcp_core.verifier import SceneStats
 
 ScriptMaterializationPolicy = str
 _SCRIPT_MATERIALIZATION_POLICIES = {"off", "auto", "require"}
@@ -966,57 +965,3 @@ def execute_with_state_digest(
 ) -> SceneDigestExecution:
     """Execute code with fail-closed before/after scene digest evidence."""
     return _script_context(context).execute_with_state_digest(code, filename=filename)
-
-
-def _scene_digest_postcondition(
-    before: SceneDigestSnapshot | Mapping[str, Any] | SceneStats | None,
-    after: SceneDigestSnapshot | Mapping[str, Any] | SceneStats | None,
-    *,
-    postcondition: Mapping[str, Any] | None,
-    verified: bool | None,
-) -> dict[str, Any] | None:
-    evidence = dict(postcondition) if postcondition is not None else {}
-    if verified is not None:
-        existing = evidence.get("verified")
-        if existing is not None and existing is not verified:
-            raise ValueError("'verified' conflicts with postcondition verification evidence")
-        evidence["verified"] = verified
-    if before is None and after is None:
-        return evidence or None
-    if before is None or after is None:
-        return ToolResultEnvelope.fail(
-            "Scene digest evidence requires both before and after observations",
-            error="scene_digest_evidence_missing",
-        ).to_dict()
-    try:
-        before_snapshot = _coerce_scene_digest_snapshot(before)
-        after_snapshot = _coerce_scene_digest_snapshot(after)
-    except SceneDigestError as exc:
-        return ToolResultEnvelope.fail(str(exc), error=exc.code).to_dict()
-    is_verified = evidence.get("verified") is True
-    if is_verified and before_snapshot.fingerprint == after_snapshot.fingerprint:
-        return ToolResultEnvelope.fail(
-            "Verified scene mutation did not change the observed scene digest",
-            error="scene_digest_postcondition_mismatch",
-        ).to_dict()
-    evidence.setdefault("verified", False)
-    evidence["scene_digest_before"] = before_snapshot.to_dict()
-    evidence["scene_digest_after"] = after_snapshot.to_dict()
-    return evidence
-
-
-def _coerce_scene_digest_snapshot(
-    value: SceneDigestSnapshot | Mapping[str, Any] | SceneStats,
-) -> SceneDigestSnapshot:
-    """Normalize an envelope snapshot while preserving fail-closed errors.
-
-    Public callers normally pass the typed snapshot returned by
-    ``execute_with_state_digest``.  Accepting the provider's ``SceneStats`` /
-    mapping shape here keeps direct adapter integrations safe: malformed or
-    tampered evidence is converted into a structured error envelope instead of
-    surfacing an ``AttributeError`` from the result builder.
-    """
-    if isinstance(value, SceneDigestSnapshot):
-        value.validate()
-        return value
-    return normalize_scene_digest(value)
